Finish case 2

+++ src/Domain/ValueObject/Tax.php
@@ -6,14 +6,14 @@ namespace Nubank\DevTest\Domain\ValueObject;
 
 class Tax
 {
-    private int $amount;
+    private float $amount;
 
-    public function __construct(int $amount)
+    public function __construct(float $amount)
     {
         $this->amount = $amount;
     }
 
-    public function getAmount(): int
+    public function getAmount(): float
     {
         return $this->amount;
     }
